freemount:  Define Frame_io_count
This corresponds to the third argument of read() or write().  It has the
same semantics, roughly speaking.  For reads, it's an upper bound on how
much data can be sent back (and can be omitted if no such limit exists).
For writes, it's an advisory notice of the total amount of data arriving
(so the receiver can preflight the request and fail early if space isn't
sufficient, or efficiently allocate contiguous space all at once).