freemount-client:  Invert SSH address semantics
When the server address is of the form host:dir//subpath, dir is used as
the Freemount root directory, and subpath is the item requested.  If the
user doesn't specify a boundary, the entire path would be considered the
root ("host:path//.").  Now, we parse host:path as "host:.//path", which
is likely to be less surprising since requests for different items share
the same root and symlinks will be treated consistently.  Also, we avoid
the infelicity of interpreting / or . as a non-directory.

This works around the issue of ssh not shell-escaping its arguments, but
only for paths without a "//" separator.  Otherwise, the root portion is
still subject to shell expansions (and the subpath remains a non-issue).