freemount-client:  Let bang indicate server path
If an SSH-form server address's path contains '!', the preceding portion
is the path to the server binary on the remote host, while the remainder
is used as the directory path (subject to splitting on "//", as before).

This allows the user to specify a custom path to a freemountd that isn't
installed conventionally (or an experimental build for testing), or even
another Freemount server altogether.