freemount-server:  Read files in a separate thread
This isn't necessary to service multiple requests from different sockets
simultaneously (provided the listener spawns separate server processes),
but it does allow GET operations to run concurrently with each other and
with other requests within the same session.  Importantly, it allows the
main thread to receive read acknowledgements so that they don't clog the
data receiver's send buffer (which would be ironic, since the whole idea
of read acknowledgements is to prevent bufferbloat).