[NEW] vraster:  Display a raster image in a window
For starters, vraster supports Mac StartupScreen format, which is merely
a copy of the screen buffer.  This version of vraster uses the Freemount
protocol over a Unix-domain socket to interact with a FORGE display, the
only working example of which is MacRelix.app.  (Classic MacRelix builds
also have FORGE, but they provide service through the filesystem instead
Freemount.)