freemount-server:  Synchronize task status access
I'm not convinced this is strictly necessary.  The real issue is whether
the task thread reaches the end before the ultimate read acknowledgement
arrives (which it almost certainly will, though we prefer to make things
bullet-proof when possible).  That said, this patch silences warnings in
helgrind, which is valuable in its own right -- better not to have false
positives obscuring any real ones.

A proper fix won't rely on read acknowledgements at all (which prohibits
using threads for non-read requests).  Rather, tasks will be reaped by a
dedicated service thread using a condition variable for synchronization.