GUI clients:  Use static storage for socket path
Otherwise, the plus::string object is destroyed before the data is used.
If the path is long enough, it will probably persist as deallocated heap
storage long enough to obscure the error, but if the path is kept in the
object's small buffer, it could be trashed in short order.  Fortunately,
that wasn't observed, but more aggressive optimization could conceivably
have that result.  Similarly, allocation-debugging utilities which spike
deallocated memory would have broken long paths.