freemount-server:  Fix race condition
The sent response informs the client that this request's ID is no longer
in use, and it's free to reuse the ID in a new request.  However, it was
possible for such a new request to arrive before the old one cleared its
status.  (In other words, check_tasks() would beat request_task::start()
in the race to lock the task's mutex.)  Defer sending the response until
the lock is acquired, so no race occurs.