interact:  Buffer images to minimize tearing
If we fill the socket's send buffer, it will contain part of the current
frame, and we'll have to wait before we can send the remainder, by which
point the author may already be painting the next frame.  To avoid this,
copy the entire frame to local memory (which has relatively low latency,
so it will finish quickly) and use that buffer to feed the socket.  This
doesn't guarantee a total absence of tearing, but it does eliminate most
of it at negligible performance cost.