interact:  Fix incorrect typecasting
The cast was wrong because it reads a single byte of data and then casts
it to uint16_t, with the data in the low byte and zero in the high byte.
The first problem is that the data byte is in the wrong location on big-
endian architectures, so interact's consumer reads the zero as a single-
byte null message before reading the $05 byte, whereupon it expects five
bytes to follow but only four arrive (eliciting a SPIEL protocol error).
The second problem is that pointer_location_buffer's device field is set
to zero on little-endian architectures.  So far this has no ill effects,
because every front end that uses this message zeroes that field anyway.